+++ tests/bench/integration.bench.ts
@@ -2,7 +2,7 @@ import { bench, describe, expect } from "vitest";
 import { routing } from "../../example/routing";
 import { Integration } from "../../src";
 
-describe("Integration", () => {
+describe.skip("Integration", () => {
   bench(
     "example",
     () => {