+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### v20.5.0
 
 - Featuring a simple profiler for the built-in logger:
-  - Introducing `BuiltinLogger::profile(msg: string)` — measures the duration until calling the returned callback;
+  - Introducing `BuiltinLogger::profile(msg: string)` — measures the duration until you invoke the returned callback;
   - Using Node Performance Hooks for measuring microtimes (less than 1ms);
   - The output severity is `debug` (will be customizable later), so logger must have the corresponding `level`;
   - It prints the duration in log using adaptive units: from picoseconds to minutes.