+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 // usage assuming that logger is an instance of BuiltinLogger
 const done = logger.profile("expensive operation");
 doExpensiveOperation();
-done(); // 2024-07-28T00:00:00.000Z debug: expensive operation '555 milliseconds'
+done(); // debug: expensive operation '555 milliseconds'
 ```
 
 ```typescript